Lots of Changes (See extended description)
1. Replacing TinyMce and SyntaxHighlighter with ckEditor and Prettify
2. Add StringExtensions
3. Code-first migrations restarted from initial
4. Blog Comment view and domain models
5. Rename SiteSettings to BlogSettings
6. IconLink Html Extension
7. PoliteCaptcha for spam filtering

+++ WR.Blog.Data/Repositories/BlogRepository.cs
@@ -28,13 +28,13 @@ namespace WR.Blog.Data.Repositories
         /// <summary>
         /// Gets the blog post by id.
         /// </summary>
-        /// <param name="id">The blog post id.</param>
+        /// <param name="blogPostId">The blog post id.</param>
         /// <returns>
         /// A blog post with the specified id.
         /// </returns>
-        public BlogPostDto GetBlogPostById(int id)
+        public BlogPostDto GetBlogPostById(int blogPostId)
         {
-            return db.BlogPosts.Include("Author").Where(b => b.Id == id).SingleOrDefault();            
+            return db.BlogPosts.Include("Author").Where(b => b.Id == blogPostId).SingleOrDefault();            
         }
 
         /// <summary>
@@ -69,12 +69,16 @@ namespace WR.Blog.Data.Repositories
         /// <summary>
         /// Deletes the blog post.
         /// </summary>
-        /// <param name="id">The id of the blog post to delete.</param>
-        public void DeleteBlogPost(int id)
+        /// <param name="blogPostId">The id of the blog post to delete.</param>
+        public void DeleteBlogPost(int blogPostId)
         {
-            BlogPostDto blogpost = db.BlogPosts.Find(id);
-            db.BlogPosts.Remove(blogpost);
-            db.SaveChanges();            
+            BlogPostDto blogpost = db.BlogPosts.Find(blogPostId);
+
+            if (blogpost != null)
+            {
+                db.BlogPosts.Remove(blogpost);
+                db.SaveChanges();             
+            }
         }
         #endregion
 
@@ -82,13 +86,13 @@ namespace WR.Blog.Data.Repositories
         /// <summary>
         /// Gets the version by id.
         /// </summary>
-        /// <param name="id">The blog post version id.</param>
+        /// <param name="blogVersionId">The blog post version id.</param>
         /// <returns>
         /// The version with the specified id.
         /// </returns>
-        public BlogVersionDto GetBlogPostVersionById(int id)
+        public BlogVersionDto GetBlogPostVersionById(int blogVersionId)
         {
-            return db.BlogVersions.Find(id);
+            return db.BlogVersions.Find(blogVersionId);
         }
 
         /// <summary>
@@ -131,43 +135,130 @@ namespace WR.Blog.Data.Repositories
         /// <summary>
         /// Deletes the blog post version.
         /// </summary>
-        /// <param name="id">The id of the version to delete.</param>
-        public void DeleteBlogPostVersion(int id)
+        /// <param name="blogVersionId">The id of the version to delete.</param>
+        public void DeleteBlogPostVersion(int blogVersionId)
         {
-            BlogVersionDto version = db.BlogVersions.Find(id);
-            db.BlogVersions.Remove(version);
-            db.SaveChanges(); 
+            BlogVersionDto version = db.BlogVersions.Find(blogVersionId);
+            
+            if (version != null)
+            {
+                db.BlogVersions.Remove(version);
+                db.SaveChanges(); 
+            } 
         }
 
         /// <summary>
         /// Deletes all versions for this blog post.
         /// </summary>
-        /// <param name="id">The blog post id.</param>
-        public void DeleteAllBlogPostVersions(int id)
+        /// <param name="blogPostId">The blog post id.</param>
+        public void DeleteAllBlogPostVersions(int blogPostId)
         {
-            db.BlogVersions.Where(v => v.VersionOf.Id == id)
+            db.BlogVersions.Where(v => v.VersionOf.Id == blogPostId)
                 .ToList().ForEach(v => db.BlogVersions.Remove(v));
             db.SaveChanges();
         }
         #endregion
 
-        #region Site Settings Methods
+        #region Blog Comment Methods
         /// <summary>
-        /// Gets the site settings.
+        /// Gets the blog comment by id.
+        /// </summary>
+        /// <param name="blogCommentId">The blog comment id.</param>
+        /// <returns>
+        /// The comment with the specified id.
+        /// </returns>
+        public BlogCommentDto GetBlogCommentById(int blogCommentId)
+        {
+            return db.BlogComments.Find(blogCommentId);
+        }
+
+        /// <summary>
+        /// Gets all blog comments.
+        /// </summary>
+        /// <returns>
+        /// Returns all blog comments
+        /// </returns>
+        public IQueryable<BlogCommentDto> GetBlogComments()
+        {
+            return db.BlogComments;
+        }
+
+        /// <summary>
+        /// Adds the blog comment.
+        /// </summary>
+        /// <param name="comment"></param>
+        /// <returns>
+        /// Returns the id of the blog comment.
+        /// </returns>
+        public int AddBlogComment(BlogCommentDto comment)
+        {
+            db.Entry(comment.BlogPost).State = EntityState.Unchanged;
+
+            if (comment.ReplyToComment != null)
+            {
+                db.Entry(comment.ReplyToComment).State = EntityState.Unchanged;
+            }
+
+            db.BlogComments.Add(comment);
+            db.SaveChanges();
+
+            return comment.Id;
+        }
+
+        /// <summary>
+        /// Updates and saves the blog comment.
+        /// </summary>
+        /// <param name="comment"></param>
+        public void UpdateBlogComment(BlogCommentDto comment)
+        {
+            db.Entry(comment).State = EntityState.Modified;
+            db.SaveChanges();
+        }
+
+        /// <summary>
+        /// Deletes the blog comment.
+        /// </summary>
+        /// <param name="blogCommentId">The id of the comment to delete.</param>
+        public void DeleteBlogComment(int blogCommentId)
+        {
+            BlogCommentDto comment = db.BlogComments.Find(blogCommentId);
+
+            if (comment != null)
+            {
+                db.BlogComments.Remove(comment);
+                db.SaveChanges();
+            } 
+        }
+
+        /// <summary>
+        /// Deletes all comments for this blog post.
+        /// </summary>
+        /// <param name="blogPostId">The blog post id.</param>
+        public void DeleteAllBlogPostComments(int blogPostId)
+        {
+            db.BlogComments.Where(c => c.BlogPost.Id == blogPostId)
+                .ToList().ForEach(c => db.BlogComments.Remove(c));
+            db.SaveChanges();
+        } 
+        #endregion
+
+        #region Blog Settings Methods
+        /// <summary>
+        /// Gets the blog settings.
         /// </summary>
         /// <returns>
         /// Returns site setttings.
         /// </returns>
-        public SiteSettingsDto GetSiteSettings()
+        public BlogSettingsDto GetBlogSettings()
         {
             return db.SiteSettings.FirstOrDefault();
         }
 
         /// <summary>
-        /// Adds the settings if they do not exist or updates site settings if they do.
+        /// Adds the settings if they do not exist or updates blog settings if they do.
         /// </summary>
         /// <param name="settings">The settings to add or update.</param>
-        public void AddOrUpdateSiteSettings(SiteSettingsDto settings)
+        public void AddOrUpdateBlogSettings(BlogSettingsDto settings)
         {
             if (settings.Id != 0)
             {
